server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-code-review"
version = "0.1.0"
description = "OpenEnv Code Review Environment for Scaler Hackathon"
authors = [
    {name = "BIJJUDAMA"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",
    "openai>=2.7.2",
    "python-dotenv",
    "openenv-core>=0.2.0",
    "orjson>=3.10.0"
]

[project.scripts]
server = "app:main"

[tool.setuptools]
py-modules = ["app"]
packages = ["graders", "tasks"]
